app/requirements.txt
blinker==1.7.0
cachetools==5.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.49.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==21.2.0
httplib2==0.22.0
hubspot-api-client==8.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.56
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.3.0
pdfkit==1.0.0
pillow==10.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.7.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.16.0
tzdata==2025.2
urllib3==2.2.1
Werkzeug==3.0.1
